third_party/blink/web_tests/platform/linux-chrome/external/wpt/svg/interact/scripted/svg-pointer-events-bbox-expected.txt

This is a testharness.js-based test.
[PASS] circle1 contains point at (36, 60)
[PASS] circle1 contains point at (42, 67)
[PASS] circle1 does not contain point at (50, 50)
[PASS] circle1 does not contain point at (50, 55)
[PASS] group1 contains point at (100, 100)
[PASS] group1 contains point at (137, 84)
[PASS] group1 contains point at (51, 156)
[PASS] group1 contains point at (70, 120)
[PASS] circle2 contains point at (400, 150)
[PASS] circle2 contains point at (432, 182)
[PASS] circle2 contains point at (361, 122)
[PASS] bbox of circle2 contains point at (438, 103)
[PASS] bbox of circle2 contains point at (450, 200)
[PASS] text1 contains point at (134, 16)
[FAIL] tspan1 contains point at (579, 46)
  assert_true: expected true got false
[PASS] text1 does not contain point at (395, 73)
[PASS] text1 does not contain point at (74, 5)
[FAIL] bbox of text1 contains point at (435, 32)
  assert_true: expected true got false
[PASS] bbox of text1 contains point at (115, 46)
[PASS] text2 contains point at (178, 146)
[FAIL] tspan2 contains point at (568, 283)
  assert_true: expected true got false
[PASS] text2 does not contain point at (319, 161)
[PASS] text2 does not contain point at (179, 131)
[PASS] bbox of text2 contains point at (295, 214)
[FAIL] bbox of text2 contains point at (444, 222)
  assert_true: expected true got false
[PASS] text3 contains point at (198, 291)
[FAIL] text3 contains point at (286, 301)
  assert_true: expected true got false
[PASS] text3 does not contain point at (302, 337)
[PASS] text3 does not contain point at (348, 335)
[PASS] image1 contains point at (19, 215)
[PASS] image1 contains point at (45, 225)
Harness: the test ran to completion.
